fixed capture discard issue
To reduce the memory footprint when benchmarking the performance of
FlueNT10G, capture data can be discarded after it has been transferred
from the FPGA board to the host system. This commit fixes an issue that
caused crashes, when multiple successive measurements are performed in
the same application and capture discard is enabled.